requirements.txt
##################################
#           PRODUCTION           #
##################################

##### Frameworks #####
autonomous-app >= 0.1.35

##### System #####
twine
wheel
setuptools
python-dotenv

##### Network #####

requests

##### Parsing #####

python-slugify
markdown

##### Misc #####

dice
